Replace `capture-details-page` with `details-page`.
