--- src/app/features/home/capture-tab/capture-details/capture-details.page.html
@@ -1,64 +1,0 @@
-<mat-toolbar color="primary" *transloco="let t">
-  <button routerLink=".." routerDirection="back" mat-icon-button>
-    <mat-icon>arrow_back</mat-icon>
-  </button>
-  <div class="toolbar-spacer"></div>
-  <button
-    (click)="openContactSelectionDialog()"
-    mat-icon-button
-    *ngIf="isRegistered$ | ngrxPush"
-  >
-    <mat-icon>send</mat-icon>
-  </button>
-  <button (click)="openOptionsMenu()" mat-icon-button>
-    <mat-icon>more_vert</mat-icon>
-  </button>
-</mat-toolbar>
-
-<div *transloco="let t" class="page-content">
-  <mat-list *ngrxLet="assetSrc$ as assetSrc" class="content">
-    <app-media
-      *ngrxLet="mimeType$ as mimeType"
-      [src]="assetSrc"
-      [mimeType]="mimeType"
-      [routerLink]="['/media-viewer', assetSrc, { mimeType: mimeType }]"
-    ></app-media>
-    <mat-list-item>
-      <mat-icon mat-list-icon>person</mat-icon>
-      <div mat-line>{{ username$ | ngrxPush }}</div>
-    </mat-list-item>
-    <mat-list-item>
-      <mat-icon mat-list-icon>place</mat-icon>
-      <div mat-line (click)="openMap()">
-        {{ location$ | ngrxPush }}
-      </div>
-    </mat-list-item>
-    <ng-container *ngrxLet="proof$ as proof">
-      <mat-list-item>
-        <mat-icon mat-list-icon>access_time</mat-icon>
-        <div mat-line>
-          {{ proof.truth.timestamp | date: 'long' }}
-        </div>
-      </mat-list-item>
-      <mat-list-item>
-        <mat-icon svgIcon="media-id-solid-black" mat-list-icon></mat-icon>
-        <div mat-line>
-          <ng-container *ngIf="isRegistered$ | ngrxPush; else notRegistered">
-            {{ proof.diaBackendAssetId }}
-          </ng-container>
-          <ng-template #notRegistered>{{ t('notRegistered') }}</ng-template>
-        </div>
-      </mat-list-item>
-    </ng-container>
-    <mat-list-item>
-      <button
-        routerLink="information"
-        class="center"
-        mat-stroked-button
-        color="primary"
-      >
-        {{ t('moreInformation') }}
-      </button>
-    </mat-list-item>
-  </mat-list>
-</div>
