--- src/app/features/home/capture-tab/capture-details/capture-details.page.scss
@@ -1,35 +1,0 @@
-.toolbar-spacer {
-  flex: 1 1 auto;
-}
-
-.page-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.content {
-  flex: 1;
-  padding-top: 0;
-  min-height: 0;
-}
-
-.end-button {
-  flex: 1;
-  flex-grow: 1;
-}
-
-button.center {
-  height: 36px;
-  width: 186px;
-  margin: auto;
-  left: 0;
-  bottom: 0;
-  color: white;
-  background-color: #564dfc;
-}
-
-.mat-list-base .mat-list-item.mat-list-item-with-avatar {
-  height: 40px;
-  padding: 0 20px;
-}
