--- src/app/features/home/capture-tab/capture-details/information/information-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { InformationPage } from './information.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: InformationPage,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class InformationPageRoutingModule {}
