--- src/app/features/home/capture-tab/capture-details/information/information.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { SharedModule } from '../../../../../shared/shared.module';
-import { InformationPageRoutingModule } from './information-routing.module';
-import { InformationPage } from './information.page';
-
-@NgModule({
-  imports: [SharedModule, InformationPageRoutingModule],
-  declarations: [InformationPage],
-})
-export class InformationPageModule {}
