--- src/app/features/home/capture-tab/capture-details/information/information.page.scss
@@ -1,5 +1,0 @@
-mat-toolbar {
-  span {
-    padding-right: 40px;
-  }
-}
