--- src/app/features/home/capture-tab/capture-details/information/information.page.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { SharedTestingModule } from '../../../../../shared/shared-testing.module';
-import { InformationPage } from './information.page';
-
-describe('InformationPage', () => {
-  let component: InformationPage;
-  let fixture: ComponentFixture<InformationPage>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [InformationPage],
-        imports: [SharedTestingModule],
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(InformationPage);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    })
-  );
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
