+++ src/app/features/home/details/details.page.html
@@ -12,7 +12,7 @@
       >
         <ion-icon name="send" slot="icon-only"></ion-icon>
       </ion-button>
-      <ion-button>
+      <ion-button (click)="openOptionsMenu()">
         <ion-icon name="ellipsis-vertical" slot="icon-only"></ion-icon>
       </ion-button>
     </ion-buttons>
@@ -57,15 +57,17 @@
             <ion-icon name="person" color="black" slot="start"></ion-icon>
             <ion-label>{{ detailedCapture.creator$ | ngrxPush }}</ion-label>
           </ion-item>
-          <ion-item>
+          <ion-item (click)="openMap()">
             <ion-icon name="location" color="black" slot="start"></ion-icon>
-            <ion-label>{{ detailedCapture.location$ | ngrxPush }}</ion-label>
+            <ion-label>{{
+              detailedCapture.locationDisplay$ | ngrxPush
+            }}</ion-label>
           </ion-item>
           <ion-item>
             <ion-icon name="time" color="black" slot="start"></ion-icon>
             <ion-label>{{ detailedCapture.timestamp$ | ngrxPush }}</ion-label>
           </ion-item>
-          <ion-item>
+          <ion-item (click)="openCertificate()">
             <ion-icon
               src="/assets/images/media-id-solid-black.svg"
               slot="start"
